Modifica classes Pessoa e Usuario para abstratas e cria Jogador
